Get Facebook profile picture
